+++ src/Rib/Server.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 
 -- | Serve generated static files with HTTP
 module Rib.Server
@@ -21,6 +22,8 @@ staticSiteServerSettings root =
     defaultSettings = defaultFileServerSettings root
 
 -- | Run a HTTP server to serve a directory of static files
+--
+-- Binds the server to host 127.0.0.1.
 serve ::
   -- | Port number to bind to
   Int ->
@@ -28,5 +31,12 @@ serve ::
   FilePath ->
   IO ()
 serve port path = do
-  putStrLn $ "[Rib] Serving at http://localhost:" <> show port
-  Warp.run port $ staticApp $ staticSiteServerSettings path
+  putStrLn $ "[Rib] Serving " <> path <> " at http://" <> host <> ":" <> show port
+  Warp.runSettings settings app
+  where
+    app = staticApp $ staticSiteServerSettings path
+    host = "127.0.0.1"
+    settings =
+      Warp.setHost (fromString host)
+        $ Warp.setPort port
+        $ Warp.defaultSettings